added html pages with formatting, updated database with additional power table, and added controller mappings with the exception of deletes

+++ SuperSightingsDB.sql
@@ -43,8 +43,8 @@ create table locations(
     locationName varchar(50),
 	`description` varchar(255),
     address varchar(100),
-    latitude INT NOT NULL,
-    longitude INT NOT NULL
+    latitude numeric(7, 5) NOT NULL,
+    longitude numeric(8, 5) NOT NULL
 );
 
 create table sightings(
